feat: refactor type imports to use new types/types module for consistency
